Fixed error with conversion to byte before mapping.
git-svn-id: http://Marc-PC/svn/Full@446 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/output-max-species-age/trunk/src/PlugIn.cs
@@ -78,7 +78,7 @@ namespace Landis.Extension.Output.MaxSpeciesAge
                     foreach (Site site in modelCore.Landscape.AllSites)
                     {
                         if (site.IsActive)
-                            pixel.MapCode.Value = (byte) SiteVars.GetMaxAge(species, (ActiveSite) site);
+                            pixel.MapCode.Value = SiteVars.GetMaxAge(species, (ActiveSite) site);
                         else
                             pixel.MapCode.Value = 0;
 
@@ -103,7 +103,7 @@ namespace Landis.Extension.Output.MaxSpeciesAge
                 foreach (Site site in modelCore.Landscape.AllSites)
                 {
                     if (site.IsActive)
-                        pixel.MapCode.Value = (byte) SiteVars.GetMaxAge((ActiveSite) site);
+                        pixel.MapCode.Value = SiteVars.GetMaxAge((ActiveSite) site);
                     else
                         pixel.MapCode.Value = 0;
